adsorption.py testing database file copied from RMG-database repo.
combination of commits with these messages:

For the database unit tests, these files need to exist here too.

New platinum adsorption groups in test_data/testing_database

Renamed the adsorptionPt data in data/test_data/thermo (WHY is this duplicated?)

It seems annoying that there is data/test_data/thermo
AND testdata/testing_database/thermo
that both need to be kept up to date.

Fixing the adsorption thermo group definitions.

So annoying this has to be fixed in the RMG-database repo
AND in TWO places here (sigh)