Change regularization dimension tracking
Now regularization is track as two lists:  the first has all "possibilities" that match at least one reaction that matches that node, the second has all "possibilities" that match all reactions that match that node.  The first is used to decide what extensions it is necessary to generate, while the second is used during regularization.