Add raise_atomtype_exception option to methods/functions that call updateAtomTypes
When using the RMG API, it can be desirable to ignore built-in RMG atom
types when building molecules from Smiles, 3D structures, etc. Often, all
we want to do is to perform isomorphism checks on the structures
involving only single bonds, so it is prohibitive to raise atom type
errors during this. Therefore, add an optional raise_atomtype_exception
argument to most methods and functions that call Molecule.updateAtomTypes
to use its already available raiseException argument.