enable subgraphIsomorphism checks between a connected molecule and a disconnected group
this is done by splitting the group, finding all of the subgraph isomorphisms between the molecule and each group fragment and determining
the combinations of those subgraph isomorphisms that do not intersect.

In order to avoid copying the group object I have added _merge and _split functions that split and merge group objects without copying the original object
this also adds some code to the end of findSubgraphIsomorphisms for putting the split group back together, since it is modified by the _split call.