Consolidate Arkane documentation input file
There was originally a lot of duplication between the input.rst
and input_pdep.rst. This creates issues with maintenance and
accuracy. This commit reduces documentation duplication by integrated
the information about pdep into the original input file and removed
the input_pdep.rst file. This should provide more accurate information
and improve maintainability.

Elaborated on Arkane's pdep sensivitity and moved the syntax example
next to the description of parameters. Added column to the parameter
list specifying whether parameter is required or not.