Inheritance for Log files Arkane
This commit makes the Log file readers for QChem, Gaussian and
Molepro inherit from the base Log class, which can reduce code
duplication going forward.

To do this, the method `determine_qm_software` was removed from
the Log class as this causes circular import errors, and the
`Log` class was moved to a separate file, in accordance with PEP-8.

Methods used in all child classes are added to the parent Log
class.

This commit also modified methods calling `determine_qm_software`,
and references to `isinstance(x,Log)` were altered to not include
the children log files.